// app/src/magi_backup_20260308_112813/data/questionnaire.types.ts
export interface IpipPersonaSeedDescriptions {
    readonly professionalDescription: string;
    readonly lifeDescription: string;
    readonly instinctNeedsDescription: string;
    readonly integratedDescription: string;
}